+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-# 0.2.0 / 2020-09-11
+# Changelog
 
-## :tada: Enhancements
+## 0.2.1 / 2020-09-11
+
+- Fix package publish
+- Fix highlight color
+
+## 0.2.0 / 2020-09-11
+
+### :tada: Enhancements
 
 - [#6](https://github.com/moshensky/pdf-visual-diff/pull/6): Allow configuration of compare-images